chore(app): move primevue options to file
